bar_conflicts: 把分类规则挪到两边都能进的那一层
`portfolio.integrity` 直接 import `market_data.bars` 闭上了一个包环
（market_data ↔ portfolio），`test_no_package_cycle_beyond_the_known_ones` 判红。
函数级 import 不是解法——那个文件自己写着「a function-level import is not a fix
for a cycle; it is a cycle that starts working」。

分类器是 foundation-shaped 的：两个 dict 进、一个词出，不碰文件、不 import 本包
任何东西，而写 refusal 的（`market_data.bars`）和把日志读回来的
（`portfolio.integrity`）都要用。按 #814 移 `sessions`/`instruments`/
`runtime_model`/`provenance` 的同一条判据放到 `clawock/bar_conflicts.py`，
`bars` 原样 re-export，调用方认识的名字不变。
